Shifted trade logic into a separate script
